+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Most recent releases are shown at the top. Each release shows:
 
 ### changed:
 - Removed duplicated item from system prompt
+- Whitelist `>/dev/null`
 
 ### fixed:
 - N/A